build: 修改版本号为0.0.5
docs: 完善注释

docs: 修改文档

feat: Android端设置、获取手势开启状态

feat: iOS端设置、获取手势开启状态

feat: Android端设置、获取Logo位置

feat: Android端设置Logo左下边距方法

feat: iOS端设置Logo左下边距方法

feat: Android端设置、获取缩放按钮位置方法

feat: iOS端设置指南针右上边距方法

feat: iOS端设置比例尺左下边距方法

feat: iOS端设置、获取是否以地图中心点缩放方法

feat: Android端设置、获取是否以地图中心点缩放方法

docs: 修改CHANGLOG.md

docs: 生成0.0.5文档

+++ lib/compass_margin.dart
@@ -8,18 +8,18 @@ class CompassMargin {
   ///
   /// [marginTop] 上边距
   CompassMargin({
-    this.marginRight = 0.0,
-    this.marginTop = 0.0,
+    this.marginRight = 0,
+    this.marginTop = 0,
   });
 
   CompassMargin.fromJson(Map<String, dynamic> json) {
-    marginRight = double.tryParse(json["marginRight"].toString()) ?? 0;
-    marginTop = double.tryParse(json["marginTop"].toString()) ?? 0;
+    marginRight = int.tryParse(json["marginRight"].toString()) ?? 0;
+    marginTop = int.tryParse(json["marginTop"].toString()) ?? 0;
   }
 
-  double marginRight = 0.0;
+  int marginRight = 0;
 
-  double marginTop = 0.0;
+  int marginTop = 0;
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = Map<String, dynamic>();
